requirements.txt
streamlit
pandas
python-binance
pytz
